+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/dao/MemberFollowerRecordDao.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface MemberFollowerRecordDao {
 
-    List<MemberFollowerRecord> queryMemberFollowerRecords(Integer mid, Integer last_count, Integer start_ts, Integer end_ts,
+    List<MemberFollowerRecord> queryMemberFollowerRecords(Long mid, Integer last_count, Integer start_ts, Integer end_ts,
                                                           Integer offset, Integer ps);
 
-    Integer queryMemberFollowerRecordsCount(Integer mid, Integer start_ts, Integer end_ts);
+    Integer queryMemberFollowerRecordsCount(Long mid, Integer start_ts, Integer end_ts);
 }
