+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/service/MemberLogService.java
@@ -6,10 +6,10 @@ import java.util.List;
 
 public interface MemberLogService {
 
-    List<MemberLog> queryMemberLogs(Integer mid, Integer start_ts, Integer end_ts,
+    List<MemberLog> queryMemberLogs(Long mid, Integer start_ts, Integer end_ts,
                                     String attr, String oldval, String newval,
                                     Integer pn, Integer ps);
 
-    Integer queryMemberLogsCount(Integer mid, Integer start_ts, Integer end_ts,
+    Integer queryMemberLogsCount(Long mid, Integer start_ts, Integer end_ts,
                                  String attr, String oldval, String newval);
 }
