+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/service/impl/MemberExServiceImpl.java
@@ -20,7 +20,7 @@ public class MemberExServiceImpl implements MemberExService {
     MemberExDao memberExDao;
 
     @Override
-    public MemberEx queryMemberByMid(Integer mid) {
+    public MemberEx queryMemberByMid(Long mid) {
         return memberExDao.queryMemberByMid(mid);
     }
