+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/service/impl/MemberTotalStatRecordServiceImpl.java
@@ -16,7 +16,7 @@ public class MemberTotalStatRecordServiceImpl implements MemberTotalStatRecordSe
     MemberTotalStatRecordDao memberTotalStatRecordDao;
 
     @Override
-    public List<MemberTotalStatRecord> queryMemberTotalStatRecords(Integer mid, Integer last_count, Integer start_ts, Integer end_ts, Integer pn, Integer ps) {
+    public List<MemberTotalStatRecord> queryMemberTotalStatRecords(Long mid, Integer last_count, Integer start_ts, Integer end_ts, Integer pn, Integer ps) {
         // pn, ps -> offset, ps
         ps = PageNumModfier.modifyPs(ps, 25000);
         pn = PageNumModfier.modifyPn(pn);
@@ -26,7 +26,7 @@ public class MemberTotalStatRecordServiceImpl implements MemberTotalStatRecordSe
     }
 
     @Override
-    public Integer queryMemberTotalStatRecordsCount(Integer mid, Integer start_ts, Integer end_ts) {
+    public Integer queryMemberTotalStatRecordsCount(Long mid, Integer start_ts, Integer end_ts) {
         return memberTotalStatRecordDao.queryMemberTotalStatRecordsCount(mid, start_ts, end_ts);
     }
 }
